+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Master
 
+## v192 (9/14/2018)
+
+* Add error messages when using unsupported Ruby versions on the Heroku 18 stack (https://github.com/heroku/heroku-buildpack-ruby/pull/809)
+
 ## v191 (8/23/2018)
 
 * Warn when `config.action_dispatch.x_sendfile_header` is set but apache and nginx are not being used (https://github.com/heroku/heroku-buildpack-ruby/pull/795)