+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Master
 
+## v193 (9/14/2018)
+
+* Fix link (https://github.com/heroku/heroku-buildpack-ruby/pull/811)
+
 ## v192 (9/14/2018)
 
 * Add error messages when using unsupported Ruby versions on the Heroku-18 stack (https://github.com/heroku/heroku-buildpack-ruby/pull/809)