+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Master
 
+## v194 (10/18/2018)
+
+* Default Ruby version is now 2.4.5 (https://github.com/heroku/heroku-buildpack-ruby/pull/821)
+
 ## v194 (10/16/2018)
 
 * Do not add the `jobs:work` command if an app does not have that rake task available (https://github.com/heroku/heroku-buildpack-ruby/pull/810)